chore: rename rehypeImgSize plugin to rehypeImg
