+++ src/pages/[...slug].tsx
@@ -6,7 +6,7 @@ import remarkGfm from "remark-gfm"
 import { join } from "path"
 
 import { getContent, getContentBySlug } from "@/lib/utils/md"
-import rehypeImgSize from "@/lib/rehype/rehypeImgSize"
+import rehypeImg from "@/lib/rehype/rehypeImg"
 import rehypeHeadingIds from "@/lib/rehype/rehypeHeadingIds"
 
 // Layouts and components
@@ -107,7 +107,7 @@ export const getStaticProps: GetStaticProps<Props, Params> = async (
       // Required since MDX v2 to compile tables (see https://mdxjs.com/migrating/v2/#gfm)
       remarkPlugins: [remarkGfm],
       rehypePlugins: [
-        [rehypeImgSize, { dir: mdDir, srcPath: mdPath, locale }],
+        [rehypeImg, { dir: mdDir, srcPath: mdPath, locale }],
         [rehypeHeadingIds],
       ],
     },
